+++ lib/model/routes.dart
@@ -1,4 +1,4 @@
 class RouteNames {
   static const String mainPage = '/';
-  static const String riverpodProviderTest = '/riverpod';
+  static const String riverpodProviderPath = '/riverpod';
 }
